+++ Cargo.toml
@@ -1,5 +1,6 @@
 [package]
 name = "meigen_bot_rust"
+default-run = "console"
 version = "0.2.2"
 authors = ["kawaemon <34652535+kawaemon@users.noreply.github.com>"]
 edition = "2018"
@@ -10,19 +11,40 @@ edition = "2018"
 anyhow = "1"
 async-trait = "0.1"
 dotenv = "0.15"
-hex = "0.4"
-mongodb = "2"
-rand = "0.8"
-regex = "1"
-ring = "0.16"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
 tokio = { version = "1", features = ["full"] }
-tokio-stream = "0.1"
+rand = "0.8"
+serde = { version = "1", features = ["derive"] }
 tracing = "0.1"
 tracing-subscriber = "0.2"
-warp = "0.3"
+
+mongodb = { version = "2", optional = true }
+regex = { version = "1", optional = true }
+tokio-stream = { version = "0.1", optional = true }
+serde_json = { version = "1", optional = true }
+ring = { version = "0.16", optional = true }
+hex = { version = "0.4", optional = true }
+warp = { version = "0.3", optional = true }
+
+[features]
+console = []
+memorydb = []
+mongodb_ = ["mongodb", "tokio-stream", "regex"]
+discord_webhook = ["warp", "hex", "ring", "serde_json"]
+
+dev = ["console", "memorydb"]
+prod = ["mongodb_", "discord_webhook"]
+default = ["prod"]
 
 [profile.release]
 lto = true
 codegen-units = 1
+
+[[bin]]
+name = "discord_webhook"
+path = "src/bin/discord_webhook.rs"
+required-features = ["discord_webhook"]
+
+[[bin]]
+name = "console"
+path = "src/bin/console.rs"
+required-features = ["console"]
