+++ rustfmt.toml
@@ -1,0 +1,3 @@
+unstable_features = true
+group_imports = "StdExternalCrate"
+imports_granularity = "Crate"
