+++ src/entrypoint/discord_webhook/mod.rs
@@ -2,20 +2,20 @@ mod interaction;
 mod model;
 mod verify;
 
-use {
-    crate::{db::MeigenDatabase, Synced},
-    anyhow::{Context, Result},
-    interaction::on_interaction,
-    serde_json::json,
-    std::{convert::Infallible, net::SocketAddr, sync::Arc},
-    tokio::sync::RwLock,
-    warp::{
-        http::StatusCode,
-        reject::Reject,
-        reply::{json as reply_json, with_status as reply_with_status, Json},
-        Filter, Rejection, Reply,
-    },
+use std::{convert::Infallible, net::SocketAddr, sync::Arc};
+
+use anyhow::{Context, Result};
+use interaction::on_interaction;
+use serde_json::json;
+use tokio::sync::RwLock;
+use warp::{
+    http::StatusCode,
+    reject::Reject,
+    reply::{json as reply_json, with_status as reply_with_status, Json},
+    Filter, Rejection, Reply,
 };
+
+use crate::{db::MeigenDatabase, Synced};
 
 // 512KB limit
 const CONTENT_LENGTH_LIMIT: u64 = 1024 * 512;
