+++ src/entrypoint/mod.rs
@@ -1,2 +1,4 @@
+#[cfg(feature = "console")]
 pub mod console;
+#[cfg(feature = "discord_webhook")]
 pub mod discord_webhook;
